requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
streamlit==1.29.0
plotly==5.17.0 